requirement.txt
aip==1.0.0.5
amqp==2.2.2
asn1crypto==0.24.0
attrs==18.1.0
Automat==0.7.0
baidu-aip==2.2.5.2
beautifulsoup4==4.6.0
billiard==3.5.0.3
cached-property==1.4.2
celery==4.1.0
certifi==2018.4.16
cffi==1.11.5
chardet==3.0.4
constantly==15.1.0
contextlib2==0.5.5
cryptography==2.2.2
cssselect==1.0.3
dateutils==0.6.6
defusedxml==0.5.0
diff-match-patch==20121119
Django==1.10.8
django-admin-easy==0.4.1
django-admin-interface==0.6.3
django-ajax-selects==1.7.1
django-celery-beat==1.1.1
django-celery-results==1.0.1
django-colorfield==0.1.14
django-constance==2.2.0
django-enumfield==1.2.1
django-flat-responsive==2.0
django-flat-theme==1.1.4
django-import-export==1.0.0
django-post-office==3.0.4
django-redis==4.9.0
django-smart-selects==1.5.2
dnspython==1.15.0
enum34==1.1.6
et-xmlfile==1.0.1
exchangelib==1.10.0
functools32==3.2.3.post2
future==0.16.0
hyperlink==18.0.0
idna==2.6
incremental==17.5.0
ipaddress==1.0.22
isodate==0.6.0
jdcal==1.4
jsonfield==2.0.2
kombu==4.1.0
lxml==4.2.1
nimbus-scrapy==1.0.0
nimbus-scrapyd-api==0.1.3
ntlm-auth==1.1.0
numpy==1.15.2
odfpy==1.3.6
openpyxl==2.5.3
pandas==0.23.4
parsel==1.5.0
pdfminer==20140328
Pillow==5.2.0
psycopg2==2.7.3
psycopg2-binary==2.7.4
pyasn1==0.4.4
pyasn1-modules==0.2.2
pycparser==2.18
PyDispatcher==2.0.5
Pygments==2.2.0
PyHamcrest==1.9.0
pykerberos==1.2.1
pymongo==3.7.1
pyOpenSSL==18.0.0
pytesseract==0.2.4
python-dateutil==2.7.3
pytz==2018.4
PyYAML==3.12
queuelib==1.5.0
raven==6.8.0
redis==2.10.6
requests==2.18.4
requests-kerberos==0.12.0
requests-ntlm==1.1.0
schedule==0.5.0
Scrapy==1.5.1
scrapy-redis==0.6.8
scrapyd-client==1.1.0
selenium==3.14.1
service-identity==17.0.0
six==1.11.0
SQLAlchemy==1.2.11
tablib==0.12.1
tesseract==0.1.3
Twisted==18.7.0
tzlocal==1.5.1
unicodecsv==0.14.1
urllib3==1.22
vine==1.1.4
w3lib==1.19.0
Wand==0.4.4
xlrd==1.1.0
XlsxWriter==1.1.1
xlwt==1.3.0
zhon==1.1.5
zope.interface==4.5.0
